Moves utility function into its own file
